+++ sitemap.xml
@@ -4,6 +4,9 @@
     <loc>https://xiaojiou176-open.github.io/apple-notes-forensics/</loc>
   </url>
   <url>
+    <loc>https://xiaojiou176-open.github.io/apple-notes-forensics/proof.html</loc>
+  </url>
+  <url>
     <loc>https://xiaojiou176-open.github.io/apple-notes-forensics/llms.txt</loc>
   </url>
   <url>
